Refactor names to standard

+++ include/SpikeMonitor.h
@@ -5,7 +5,7 @@
 
 class SpikeMonitor {
 public:
-    void on_spike(double time, int neuron_id);
+    void on_spike(double time, size_t neuron_id);
     void reset_spikes();
 
     // Public for direct access from python
