+++ src/SpikeMonitor.cpp
@@ -1,6 +1,6 @@
 #include "SpikeMonitor.h"
 
-void SpikeMonitor::on_spike(double time, int neuron_id) {
+void SpikeMonitor::on_spike(double time, size_t neuron_id) {
     this->spike_list.emplace_back(time, neuron_id);
 }
